requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
langchain-huggingface>=0.0.3
langchain-chroma>=0.1.0
chromadb>=0.4.24
youtube-transcript-api>=0.6.0
huggingface_hub>=0.21.4
python-dotenv>=1.0.0
tqdm>=4.66.2
